backend/requirements.txt
# FastAPI и веб-сервер
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# База данных
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# AI модели
openai==1.10.0              # GPT-4o for production

# Аутентификация
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Google OAuth & API
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.147.0  # For Google Sheets API (formula executor)

# Утилиты
python-dotenv==1.0.0
httpx==0.26.0
# EMERGENCY FIX: 10 Nov 2025 15:00-v1.5.0 - Railway caching issue - GPT-4o ONLY
